+++ docs/RELEASE-NOTES.md
@@ -25,7 +25,7 @@ This is not a release yet.
 
 * #556 Fixed Missing parentheses in `SQL` for disjunctions when temporary tables are not used by the `SQLStore`
 * #764 Fixed DB error when a `#ask` query contains `order=random` for a `sqlite` or `postgres` DB platform (disabled `smwgQRandSortingSupport` for `postgres`)
-* #860 Fixed escape character usage in `SPARQLStore`, `SQLStore` 
+* #860 Fixed escape character usage in `SPARQLStore`, `SQLStore`
 * #860 Fixed handling of an empty result send by the `SPARQLStore` Sesame connector
 * #861 Fixed owl property export declaration
 * #863 Fixed missing interwiki encoding for the RDF export